# web/api-doc/swagger.yaml
openapi: 3.0.0
info:
  title: 'Web Semester 6'
  version: '0.1'
...
paths:
  /site/login:
    post:
      operationId: 0e6ee86b2507b85475c45d4fe785210a
      tags:
        - Site Controller
      parameters:
        -
          name: username
          in: formData
          description: 'Имя пользователя'
          required: true
          type : 'string'
        -
          name: password
          in: formData
          description: 'Пароль'
          required: true
          type: 'string'
        -
          name: rememberMe
          in: formData
          description: 'Запомнить меня'
          required: true
          type: 'bool'
      responses:
          '400':
            description: 'Неверные введенные данные'
          '200':
            description: 'Успешно залогинились'
...